fix(leaderboard): 统一未知类型运行态契约
- 由公开策略统一判定敏感与未定义排行榜类型\n- 补齐 Gateway PC / mobile 与登录态成组验收记录\n- 关闭 F4-S 并同步当前规划和开发日志

+++ Radish.Api.Tests/Controllers/LeaderboardControllerTest.cs
@@ -24,7 +24,7 @@ public sealed class LeaderboardControllerTest
         var service = new Mock<ILeaderboardService>(MockBehavior.Strict);
         var controller = CreateController(service.Object);
 
-        var result = await controller.GetLeaderboard(type);
+        var result = await controller.GetLeaderboard((int)type);
 
         Assert.False(result.IsSuccess);
         Assert.Equal(400, result.StatusCode);
@@ -46,7 +46,7 @@ public sealed class LeaderboardControllerTest
         var service = new Mock<ILeaderboardService>(MockBehavior.Strict);
         var controller = CreateController(service.Object, currentUserId: 1001);
 
-        var result = await controller.GetMyRank(type);
+        var result = await controller.GetMyRank((int)type);
 
         Assert.False(result.IsSuccess);
         Assert.Equal(400, result.StatusCode);
@@ -63,7 +63,7 @@ public sealed class LeaderboardControllerTest
         var service = new Mock<ILeaderboardService>(MockBehavior.Strict);
         var controller = CreateController(service.Object, currentUserId: 1001);
 
-        var result = await controller.GetMyRank(LeaderboardType.HotProduct);
+        var result = await controller.GetMyRank((int)LeaderboardType.HotProduct);
 
         Assert.False(result.IsSuccess);
         Assert.Equal(400, result.StatusCode);
